Cast attachment quota sum to BIGINT
SUM(bigint) returns NUMERIC in Postgres, which fails to decode as i64
and turned every attachment upload into a 500.

+++ crates/backend/src/main.rs
@@ -1531,7 +1531,8 @@ async fn store_attachments(
         .execute(&mut *tx)
         .await?;
     let (used_bytes,): (i64,) = sqlx::query_as(
-        "SELECT COALESCE(SUM(a.size_bytes), 0) \
+        // SUM(bigint) yields NUMERIC in Postgres; cast back so it decodes as i64.
+        "SELECT COALESCE(SUM(a.size_bytes), 0)::BIGINT \
          FROM attachments a \
          JOIN tasks t ON t.id = a.task_id \
          JOIN projects p ON p.id = t.project_id \
